The AdminController and admin view were created.
The AdminController was created
The index method was created
The admin\index.blade.php file was created by copying the dashboard.blade.php the directive changed to x-admin-layout
The admin.blade.php file was created inside the layouts folder by duplicating the app.blade.php
The AdminLayout file was created inside the app\View folder by duplicating the AppLayout.php file. The name of the class was changed to AdminLayout and also render method was changed.
The admin.index route was created