create page link, create function and create page
The link was create in the index page to navigate to the create page.
The links was added to the other files too.
The create.blade.php files were created for each objects by duplicating the index file in the same directory and remove the table syntax and the copied syntax was added.
Some unwanted div elements were removed from the template.
The create function were implemented in each controller files