Form validation, create and retrive data from DB
Fillable properties were added in the Category.php model.
CategoryStoreRequest.php was created with the artisan command.
CategoryStoreRequest was implemented with the syntax
The return value of the authorise method was change to true.
The validation rules for form fields were added in the rules method.
Then the store method of the CategoryController was implemented with the help of the CategoryStoreRequest.
The image was stored into the `public/categories` directory. The data row was created and the page redirected to the index page.
The data was retrived and showed in the index page.