Create menu functionality
The form fields were added

method, action were added

error message were added

Name attributes was added as array to the select tag
The return value of the authorise method was changed to true.

The validation rules were added
The fillable properties were added in the Menu model.
The store method was implemented in the Admin\MenuController.php